More update. Add outbox patttern. Add persistence pattern. Add more check in application layer. Add update all entity in application layer.

+++ crs/Services/Book/Application/Carts/Commands/DeleteItemInCart/DeleteItemInCartCommandHandler.cs
@@ -31,8 +31,14 @@ internal sealed class DeleteItemInCartCommandHandler(
         var item = await _cartItemRepository.GetAsync(itemId, cancellationToken);
         var book = await _bookRepository.GetAsync(item.Book.Id, cancellationToken);
 
-        var bookQuantity = QuantityBook.Create(item.Quantity.Value).Value;
-        book.AddQuantity(bookQuantity);
+        var bookQuantityResult = QuantityBook.Create(item.Quantity.Value);
+
+        if (bookQuantityResult.IsFailure)
+        {
+            return bookQuantityResult;
+        }
+
+        book.AddQuantity(bookQuantityResult.Value);
 
         await _cartItemRepository.DeleteAsync(itemId, cancellationToken);
         await _bookRepository.UpdateAsync(book, cancellationToken);
